Doing review suggestions
