+++ devtools/dev-requirements_conda.yml
@@ -22,6 +22,7 @@ dependencies:
     - orthax
     - plotly >= 5.16, < 6.0
     - pylatexenc >= 2.0, < 3.0
+    - quadax >= 0.2.2
     # building the docs
     - sphinx-github-style >= 1.0, < 2.0
     # testing and benchmarking
