+++ requirements.txt
@@ -12,5 +12,6 @@ orthax
 plotly >= 5.16, < 6.0
 psutil
 pylatexenc >= 2.0, < 3.0
+quadax >= 0.2.2
 scipy >= 1.7.0, < 2.0.0
 termcolor
