+++ requirements_conda.yml
@@ -21,3 +21,4 @@ dependencies:
     - orthax
     - plotly >= 5.16, < 6.0
     - pylatexenc >= 2.0, < 3.0
+    - quadax >= 0.2.2
